Two major changes
1. Now the events system works with new cool functionalities;
- The modules can listen to events.
- Any thing with access to the modules context can broadcast  its own
custom events specially modules ;)
2. Now the modules must return the results of what they've done, (now
you must do "return $this->display("myView")" instead of simply
"$this->display("myView")"). This is due to the need of allowing another
parts of the system to modify the final output or to have modules that,
when invoked, returns something that is not suposed to be rendered, an
extensibility improvement indeed.