Some urgent optimization
The new feature that allows modules to listen to events became a huge
bottleneck, due to the extensive usage of annotations, but doctrine
already provided an optimization way, the one I used is the CachedReader
in combination with a FileCache. This cache system wil be used until I
figure out how to detect the different cache alternatives and decide
which is better if available.

By the way, I tried a bootstrapping mechanism similar to the symfony's
bootstrap huge file method, but it resulted quite slower! I suspect that
is due to the usage of doc-block annotations.

P.D.: fixed a huge mistake with the EntryPointInterface which I
erroneusly wrote as a class... but PHP didn'd alert me about :|